Fix loose/unpacked object reading
Loose objects are a bit of a pain as they're compressed as a whole file where
as packed objects are compressed only as byte segments in the pack file.

I resolved this discrepancy with a delay that uncompresses the data in the
appropriate way but only if needed. This may make things more complex later
when trying to reuse java.zip.Inflater objects for performance.

We should now be handling both loose and packed objects properly and, for the
time being, are not handling objects in MIDX (multi-pack-index) files.